refactor: bot related fns -> bot.js
